refactor: move transformByType to generic utils file
