+++ src/utils/system/__tests__/tree-walker.spec.js
@@ -1,18 +1,4 @@
-import { createWalkerModule, transformByType } from '../tree-walker';
-
-it('routes based on type', () => {
-  const routes = {
-    a: (x) => ['a', x],
-    b: (x) => ['b', x]
-  };
-
-  expect(transformByType({ type: 'a' }, routes)).toEqual(
-    ['a', { type: 'a' }]
-  );
-  expect(transformByType({ type: 'b' }, routes)).toEqual(
-    ['b', { type: 'b' }]
-  );
-});
+import { createWalkerModule } from '../tree-walker';
 
 /*
   type Node = Branch | Leaf;
