+++ src/utils/system/tree-walker.ts
@@ -1,29 +1,5 @@
+import { Typed } from '../nodes';
 import { Module } from './system';
-
-type Typed<Ts extends string = string> = {
-  type: Ts;
-};
-
-type Routes<N extends Typed, Args extends unknown[], R> = {
-  [T in N['type']]: (n: Extract<N, Typed<T>>, ...args: Args) => R
-};
-
-type TransformByType = {
-  <N extends Typed, Args extends [] | unknown[], R>(n: N, args: Args, routes: Routes<N, Args, R>): R;
-  <N extends Typed, R>(n: N, routes: Routes<N, [], R>): R;
-};
-
-export const transformByType: TransformByType = <N extends Typed, Args extends unknown[], R>(
-  n: N,
-  _args: Args | Routes<N, [], R>,
-  _routes?: Routes<N, Args, R>
-): R => {
-  const routes = (_routes == null ? _args : _routes) as Routes<N, Args, R>;
-  const args = (_routes == null ? [] : _args) as Args;
-
-  const matching = routes[n.type as N['type']];
-  return matching(n as Extract<N, Typed<N['type']>>, ...args);
-};
 
 
 export type Walk<T extends Typed, R> = (x: T) => R;
